Done gata tot, fara bill

+++ peroject3/src/main/java/model/Purchase.java
@@ -20,7 +20,7 @@ public class Purchase {
     public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
-    public double getTotalPrice() {
+    public int getTotalPrice() {
         return totalPrice;
     }
     public void setTotalPrice(int totalPrice) {
